Set up the test suite with a consistent registry
